* 'ready' of https://github.com/WeNeedName/WeNeedName_backend: (21 commits)
  일정 response crop 수정
  수확량 정말 마지막 수정 제발
  마지막의 마지막 수확량 수정
  진짜진짜 마지막 제발 수확량 수정
  수확량 통계 마지막 수정
  지출,수확량 통계 수정
  수확량,수입지출 통계 오류 수정
  오류수정하는데 이거 깜빡함
  작업시간 오류 수정
  작업시간 오류 수정
  jenkins gradlew test 원인 출력
  그럼 이려면은 되겠네?
  test의 test
  test
  test 원상복귀..
  jenkins test 실험
  원형차트 문자열 to 숫자
  장부 1달치 조회 버그 수정
  장부, 일정 월단위 조회 범위 수정
  임시 jenkins test 수정
  ...

# Conflicts:
#	src/main/java/com/example/formproject/repository/WorkLogRepository.java
#	src/main/java/com/example/formproject/service/WorkLogService.java

+++ src/main/java/com/example/formproject/dto/response/LineChartDataDto.java
@@ -19,9 +19,9 @@ public class LineChartDataDto {
 
     @Builder.Default
     @Schema(type = "List",example = "[100,200]")
-    private List<Long> data = new ArrayList<>();
+    private List<Integer> data = new ArrayList<>();
 
-    public void addData(long data){
+    public void addData(int data){
         this.data.add(data);
     }
 }
